RAG-sqlite/pyproject.toml
[project]
name = "rag-sqlite"
version = "0.1.0"
description = ""
authors = [
    { name = "Priyanshu-K07", email = "priyanshukapadnis07@gmail.com" }
]
requires-python = ">=3.12,<4.0"

dependencies = [
    "langchain",                 # Pin base version
    "langchain-community",              # Let Poetry choose compatible versions
    "langchain-core",
    "langchain-google-genai",
    "langchain-experimental",
    "chromadb",
    "python-dotenv"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
